+++ src/pages/feed.tsx
@@ -1,15 +1,48 @@
 import { Box } from "../utilities/material-ui/material-components";
 import {
-  DiscussionCardsSection,
+  QuestionCardsSection,
   TrendingSection,
   AddPostMobile,
   Sidenav,
 } from "../components/index";
 import React from "react";
+import { useAppDispatch, useAppSelector } from "../redux/customHook";
+import { getQuestions } from "../redux/slices/feedSlice";
+import { RootState } from "../redux/store";
+import { Question } from "../types/Question";
 
 export function Feed() {
   const [open, setOpen] = React.useState(false);
   const handleOpen = () => setOpen(true);
+  const dispatch = useAppDispatch();
+  const { questions } = useAppSelector((state: RootState) => state.feed);
+  const { userDetails } = useAppSelector(
+    (state: RootState) => state.authentication
+  );
+  const [questionsToDisplay, setQuestionsToDisplay] = React.useState<
+    Question[]
+  >([]);
+
+  React.useEffect(() => {
+    if (questions.length === 0) {
+      dispatch(getQuestions());
+    }
+    console.log("here");
+  }, []);
+  React.useEffect(() => {
+    console.log(
+      questions?.filter(
+        (ques: Question) => userDetails.following.includes(ques.username) ?? []
+      ),
+      userDetails.following
+    );
+
+    setQuestionsToDisplay((value: Question[]) =>
+      questions?.filter((ques: Question) =>
+        userDetails.following.includes(ques.username)
+      )
+    );
+  }, [questions]);
   return (
     <Box
       sx={{
@@ -23,7 +56,10 @@ export function Feed() {
       component="main"
     >
       <Sidenav handleOpen={handleOpen} open={open} setOpen={setOpen} />
-      <DiscussionCardsSection title="Latest Posts" />
+      <QuestionCardsSection
+        title="Latest Posts"
+        questions={questionsToDisplay}
+      />
       <TrendingSection />
       <AddPostMobile handleOpen={handleOpen} />
     </Box>
