+++ src/types/Comment.ts
@@ -1,0 +1,7 @@
+export interface Comment {
+  _id: string;
+  username: string;
+  firstName: string;
+  lastName: string;
+  commentText: string;
+}
