--- src/types/Discussion.ts
@@ -1,9 +1,0 @@
-export interface Discussion{
-    id: number;
-    title: string;
-    description: string;
-    postedBy: string;
-    votes: number;
-    tags: string[];
-    time: number;
-}
