+++ src/types/Vote.ts
@@ -1,0 +1,4 @@
+export interface Vote {
+  upvotedBy: string[];
+  downvotedBy: string[];
+}
